@@ -3,7 +3,8 @@ import { AlarmState } from "./alarmHistory";
 import { calculateMetric } from "./common";
 
 export async function handler(event: SNSEvent) {
-  const metric = calculateMetric('MTTF', AlarmState.ALARM, AlarmState.OK)
+
+  const metric = calculateMetric('MTBF', AlarmState.ALARM, AlarmState.ALARM)
   
   return await metric(event)
 
