@@ -3,8 +3,9 @@ import { AlarmState } from "./alarmHistory";
 import { calculateMetric } from "./common";
 
 export async function handler(event: SNSEvent) {
-  const metric = calculateMetric('MTTF', AlarmState.ALARM, AlarmState.OK)
-  
+
+  const metric = calculateMetric('MTTR', AlarmState.OK, AlarmState.ALARM)
+
   return await metric(event)
 
 }
